hardware/requirements.txt
# Anamnesis.fm Physical Radio - Python Dependencies

# OLED Display
luma.oled==3.13.0
luma.core==2.4.2
pillow>=10.0.0

# GPIO and hardware
RPi.GPIO==0.7.1
spidev==3.6
adafruit-circuitpython-mcp3xxx==1.4.14
adafruit-blinka==8.25.0

# HTTP client
requests>=2.31.0
urllib3>=2.0.0

# Audio
python-mpv==1.0.6

# Async support
asyncio-throttle==1.0.2
